// colaboradoresAngular/src/app/shared/interfaces/initiative-interface.ts
export class InitiativeModel
{
  idIniciativa:  any;
  folioIniciativaAsociada:  string;
  iniciativaEstrategica:  string;
  descripcion: string;
  folio:  string;
  numIniciativa:  string;
  tituloIniciativa:  string;
  idSolicitante:  string;
  tituloSolicitante:  string;
  idEstatus:  string;
  tituloEstatus:  string;
  idModOp:  string;
  tituloModOp:  string;
  porcentajeAvance:  string;
  fechaInicio:  string;
  fechaFin:  string;
  timeReport:  string;
  numIntegrantes:  string;
  idTipoProyecto:  string;
  tituloTipoProyecto:  string;
  idDireccion:  string;
  tituloDireccion:  string;
  iniciativaAsociada:  string;
  contador: string;

}